refactor(mcp): replace router switch with route registry
