+++ TablePro/Core/MCP/MCPRouteHandler.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+protocol MCPRouteHandler: Sendable {
+    var methods: [HTTPRequest.Method] { get }
+    var path: String { get }
+    func handle(_ request: HTTPRequest) async -> MCPRouter.RouteResult
+}
